Fix [UI] project request is sent 3 times (#2901)
